day 16 - part 2 - new version based on python code
my very first idea was better, but I fell to some random suggestion
online (create a graph of valid openable valves). It worked for the
first part, but not for the second.

the abstraction was too insane to work on decent time. I've found a
python version on Reddit which was too dang small to be true. But it
was!

since it was clear to me my solution was crap for part two, I decided to
blow it up and rewrite. Ironically, the python code was used to
re-validate my answers (since I can't double-submit part 1).

eventually, the challenge of reverse-engineer the python code was too
much. I took that as a better usage of my time, with better yields on
learning lessons.

after I understood how that python worked, I wrote my own version. And
it is incredibly dang similar. Too bad I don't have the original link,
but it was definitely better for myself understand that incredible
solution than following random people online with solutions that works
after 40 minutes. Or spending too much time ingesting Ibuprofen to
counter the migraines my own solution was yielding me.